Fixes 1 - 5 implemented with full STDOUT support
All changes from August 15 till August 21. All fixes are except Fix_2, which is just to improve performance. Fixed the following: i) Started at the wrong vertex (read end edge end instead of read start edge start) for the conjugate Dijkstra graph. ii) Improper calculation of the reduction in barcode distance, reducing it by the length of the read instead of the distance between the read start/end to the edge start/end. iii) Improper checking of the edge that the first read ends and the second read starts on. iv) Did not check the reverse complement of the read (i.e.: did not check the reverse complement graph of either the forward or the conjugate).